fix nav logic & payment css
